requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2024.1
dj-database-url==2.2.0
Django==5.0.4
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-crispy-forms==2.1
django-crontab==0.7.1
django-embed-video==1.4.4
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-mail-admin==0.3.2
django-phonenumber-field==7.3.0
django-social-share==2.3.0
django-taggit==5.0.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
jsonfield==3.1.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
setuptools==69.5.1
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.0
whitenoise==6.6.0
